refactor: reorganize package structure for Cluster and Environment classes
