+++ src/main/java/org/qubership/colly/ClusterResourcesLoader.java
@@ -14,14 +14,14 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.qubership.colly.cloudpassport.CloudPassport;
 import org.qubership.colly.cloudpassport.CloudPassportEnvironment;
 import org.qubership.colly.cloudpassport.CloudPassportNamespace;
-import org.qubership.colly.db.Cluster;
-import org.qubership.colly.db.Environment;
-import org.qubership.colly.db.EnvironmentType;
-import org.qubership.colly.db.Namespace;
+import org.qubership.colly.db.data.Cluster;
+import org.qubership.colly.db.data.Environment;
+import org.qubership.colly.db.data.EnvironmentType;
+import org.qubership.colly.db.data.Namespace;
 import org.qubership.colly.monitoring.MonitoringService;
-import org.qubership.colly.storage.ClusterRepository;
-import org.qubership.colly.storage.EnvironmentRepository;
-import org.qubership.colly.storage.NamespaceRepository;
+import org.qubership.colly.db.ClusterRepository;
+import org.qubership.colly.db.EnvironmentRepository;
+import org.qubership.colly.db.NamespaceRepository;
 
 import java.io.IOException;
 import java.net.URI;
