+++ src/main/java/org/qubership/colly/db/NamespaceRepository.java
@@ -1,10 +1,9 @@
-package org.qubership.colly.storage;
+package org.qubership.colly.db;
 
 import io.quarkus.hibernate.orm.panache.PanacheRepository;
 import jakarta.enterprise.context.ApplicationScoped;
-import org.qubership.colly.db.Cluster;
-import org.qubership.colly.db.Environment;
-import org.qubership.colly.db.Namespace;
+import org.qubership.colly.db.data.Cluster;
+import org.qubership.colly.db.data.Namespace;
 
 import java.util.List;
