+++ src/test/java/org/qubership/colly/ClusterResourcesLoaderTest.java
@@ -7,7 +7,6 @@ import io.kubernetes.client.openapi.models.*;
 import io.quarkiverse.wiremock.devservice.ConnectWireMock;
 import io.quarkiverse.wiremock.devservice.WireMockConfigKey;
 import io.quarkus.test.TestTransaction;
-import io.quarkus.test.component.TestConfigProperty;
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.inject.Inject;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
@@ -17,11 +16,11 @@ import org.junit.jupiter.api.Test;
 import org.qubership.colly.cloudpassport.CloudPassport;
 import org.qubership.colly.cloudpassport.CloudPassportEnvironment;
 import org.qubership.colly.cloudpassport.CloudPassportNamespace;
-import org.qubership.colly.db.Environment;
-import org.qubership.colly.db.EnvironmentType;
-import org.qubership.colly.db.Namespace;
-import org.qubership.colly.storage.EnvironmentRepository;
-import org.qubership.colly.storage.NamespaceRepository;
+import org.qubership.colly.db.data.Environment;
+import org.qubership.colly.db.data.EnvironmentType;
+import org.qubership.colly.db.data.Namespace;
+import org.qubership.colly.db.EnvironmentRepository;
+import org.qubership.colly.db.NamespaceRepository;
 
 import java.net.URI;
 import java.util.List;
